refactor: add navigation fix exit message
